refactored directory name for error responses
